+++ Proyecto_Fin_Hibrido/Controllers/NominasController.cs
@@ -21,15 +21,45 @@ namespace Proyecto_Fin_Hibrido.Controllers
     {
         private Proyecto_Fin_HibridoEntities db = new Proyecto_Fin_HibridoEntities();
 
-        // GET: api/Nomina
+        // GET: api/Nominas
+        [Route("api/Nominas")]
         public IQueryable<Nomina> GetNomina()
         {
             Request.Properties["Count"] = db.Nomina.Count();
             return db.Nomina;
         }
 
-        // GET: api/Nomina/5
+        // GET: api/Nominas?id[0]=1&id[1]=2
+        [Route("api/Nominas/{id?}")]
+        public IQueryable<Nomina> GetNomina([FromUri] int[]id)
+        {
+
+            if (id == null)
+            {
+                Request.Properties["Count"] = db.Nomina.Count();
+                return db.Nomina;
+            }
+            else
+            {
+
+                
+                List<Nomina> res = new List<Nomina>();
+                foreach (int uid in id)
+                {
+                    res.Add(db.Nomina.Find(uid));
+                }
+                Request.Properties["Count"] = res.Count();
+                return res.AsQueryable<Nomina>();
+            }
+
+        }
+
+
+
+
+        // GET: api/Nominas/5
         [ResponseType(typeof(Nomina))]
+        [Route("api/Nominas/{id}")]
         public IHttpActionResult GetNomina(int id)
         {
             Nomina nomina = db.Nomina.Find(id);
@@ -41,8 +71,9 @@ namespace Proyecto_Fin_Hibrido.Controllers
             return Ok(nomina);
         }
 
-        // PUT: api/Nomina/5
+        // PUT: api/Nominas/5
         [ResponseType(typeof(Nomina))]
+        [Route("api/Nominas/{id}")]
         public IHttpActionResult PutNomina(int id, Nomina nomina)
         {
             if (!ModelState.IsValid)
@@ -76,8 +107,9 @@ namespace Proyecto_Fin_Hibrido.Controllers
             return Ok(nomina);
         }
 
-        // POST: api/Nomina
+        // POST: api/Nominas
         [ResponseType(typeof(Nomina))]
+        [Route("api/Nominas")]
         public IHttpActionResult PostNomina(Nomina nomina)
         {
             if (!ModelState.IsValid)
@@ -91,8 +123,9 @@ namespace Proyecto_Fin_Hibrido.Controllers
             return Ok(nomina);
         }
 
-        // DELETE: api/Nomina/5
+        // DELETE: api/Nominas/5
         [ResponseType(typeof(Nomina))]
+        [Route("api/Nominas/{id}")]
         public IHttpActionResult DeleteNomina(int id)
         {
             Nomina nomina = db.Nomina.Find(id);
@@ -101,6 +134,10 @@ namespace Proyecto_Fin_Hibrido.Controllers
                 return NotFound();
             }
 
+            if(db.Empleado.Count(e => e.IdNomina == id) > 0)
+            {
+                return Conflict();
+            }
             db.Nomina.Remove(nomina);
             db.SaveChanges();
